requirement.txt
fastapi
uvicorn
pdfplumber
python-docx
pytesseract
pillow 
pandas 
openpyxl 
python-multipart 
python-magic 
jinja2pip install 

